手机验证码使用session时与前端沟通异常，改为使用redis缓存，同时修改实名认证的相关session功能。 修复图片数组接收异常的bug，修改相关路径与思路。 修复充值和提现金额为小数异常的问题。 纠正部分注释。

+++ project/api/users/user_money.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from flask import make_response, jsonify, request
 from flask_restful import Resource, reqparse
 
@@ -11,7 +13,7 @@ class User_money(Resource):  # 修改用户金额
     @check_status([0, 1, 3])
     def put(self):
         parser = reqparse.RequestParser()
-        parser.add_argument('money', type=int, required=True, help='请输入金额')
+        parser.add_argument('money', type=float, required=True, help='请输入金额')
         parser.add_argument('type', type=str, required=True, help='请输入修改类型')
         args = parser.parse_args()
         if args['type'] not in ('recharge', 'withdrawal'):
@@ -22,13 +24,13 @@ class User_money(Resource):  # 修改用户金额
         if args['type'] == 'recharge':
             if args['money'] < 0:
                 return make_response(jsonify(code=400, message='充值金额不能为负数'), 400)
-            user.money += args['money']
+            user.money += Decimal(args['money'])
             db.session.commit()
             logger.debug(f'用户{user.username}充值成功')
             return make_response(jsonify(code=201, message='充值成功', data={'money': user.money}), 201)
         if args['type'] == 'withdrawal':
             if user.money >= args['money']:
-                user.money -= args['money']
+                user.money -= Decimal(args['money'])
                 db.session.commit()
                 logger.debug(f'用户{user.username}提现成功')
                 return make_response(jsonify(code=201, message='提现成功', data={'money': user.money}), 201)
